// src/app/models/statisticalHomeModel.ts
export interface StatisticalHome {
    budgetedAccumSalessPD?: any;
    budgetedDailySales?: any;
    userName?: any;
    actualDailySales?: any;
    lessVAT10?: any;
    lessSC5?: any;
    estimatedDailySales?: any;
    estimatedAccumSalesPd?: any;
    variancedaily?: any;
    varianceAccumulated?: any;
    customerCount?: any;
    checkAverage?: any;
    wastageVND?: any;
    wastagePercent?: any;
    actualStoreWorkingHours?: any;
    payrollVND?: any;
    payrollPercent?: any;
    budgetPayrollPercent?: any;
    actualSPMH?: any;
    hoursShouldHaveUsded?: any;
    dayName?: any;
    dayMonth?: any;
    lessMgntMeab?: any;
    variance?: any;
    actualDailySalesUSD?: number,
    lessMgntMeabUSD?: number,
}